# openapi.yaml
openapi: 3.1.0
info:
  title: Movie Maven API
  version: 1.1.0
  summary: Catalog and ingest movies from TMDB into the Movie Maven store.
  description: |
    HTTP API for the Movie Maven backend. The server currently listens on
    `http://localhost:8001` and returns JSON.

    Populate endpoints fetch from [The Movie Database (TMDB)](https://developer.themoviedb.org/)
    Discover API, upsert by `tmdb_id`, and persist to PostgreSQL. Listing reads
    the local catalog only.

    This OpenAPI file is the integration contract. Import it into Postman,
    Insomnia, Stoplight, or OpenAPI code generators.
  contact:
    name: Movie Maven
servers:
  - url: http://localhost:8001
    description: Local development
tags:
  - name: catalog
    description: Read the persisted movie catalog.
  - name: ingest
    description: Pull movies from TMDB and upsert them into the catalog.
paths:
  /movies:
    get:
      tags: [catalog]
      operationId: listMovies
      summary: List cataloged movies
      description: |
        Returns a page of non-deleted movies from the local database, newest
        `release_date` first (`NULLS LAST`).

        Defaults to `page=1` and `page_size=20`. `page_size` may be at most 100.
        A page past the last page returns an empty `movies` array.

        Pass `genre_id` to only return movies whose `genre_ids` include that
        TMDB movie genre id. Genre ids can be found via `GET /populate-genres`.
      parameters:
        - name: page
          in: query
          required: false
          description: 1-based page number. Defaults to 1.
          schema:
            type: integer
            minimum: 1
            default: 1
            example: 1
        - name: page_size
          in: query
          required: false
          description: Number of movies per page. Defaults to 20, maximum 100.
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
        - name: genre_id
          in: query
          required: false
          description: >-
            TMDB movie genre id to filter by. Only movies whose `genre_ids`
            array contains this id are returned. Omit to return all genres.
          schema:
            type: integer
            example: 35
      responses:
        "200":
          description: One page of catalog movies.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListMoviesResponse"
        "400":
          description: Invalid `page`, `page_size`, or `genre_id` query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Failed to load movies from the database.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /genres:
    get:
      tags: [catalog]
      operationId: listGenres
      summary: List cataloged genres
      description: |
        Returns all genres from the local database, ordered by `media_type`
        then `name`.

        Pass `media_type` to only return genres from that TMDB list. Genres
        are populated via `GET /populate-genres`.
      parameters:
        - name: media_type
          in: query
          required: false
          description: Restrict results to this TMDB genre list. Omit to return both.
          schema:
            type: string
            enum: [movie, tv]
            example: movie
      responses:
        "200":
          description: All cataloged genres, optionally filtered by media_type.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListGenresResponse"
        "400":
          description: Invalid `media_type` query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Failed to load genres from the database.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  /populate-genres:
    get:
      tags: [ingest]
      operationId: populateGenres
      summary: Ingest TMDB's full movie and TV genre lists
      description: |
        Fetches TMDB's movie genre list (`/genre/movie/list`) and TV genre
        list (`/genre/tv/list`) and upserts both into the local catalog,
        keyed by `id` and `media_type`.
      responses:
        "200":
          description: Genres saved (insert or update).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PopulateGenresResponse"
        "502":
          description: TMDB request failed before a response body was received.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: TMDB payload could not be parsed or genres could not be saved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: |
            TMDB returned a non-200 status. The status code and JSON body are
            forwarded from TMDB unchanged.

  /populate-movies:
    get:
      tags: [ingest]
      operationId: populateMovies
      summary: Ingest one page of 2025 English-language movies
      description: |
        Fetches a single TMDB Discover page for primary release year 2025,
        original language `en`, `vote_count >= 50`, sorted by
        `primary_release_date.desc`. Matching movies are upserted on `tmdb_id`.
      responses:
        "200":
          description: Movies saved (insert or update).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PopulateMoviesResponse"
        "502":
          description: TMDB request failed before a response body was received.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: TMDB payload could not be parsed or movies could not be saved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        default:
          description: |
            TMDB returned a non-200 status. The status code and JSON body are
            forwarded from TMDB unchanged.

  /populate-movies-since:
    get:
      tags: [ingest]
      operationId: populateMoviesSince
      summary: Ingest movies whose primary release is on or after a given year
      description: |
        Fetches TMDB Discover pages for English-language movies with
        `vote_count >= 200`, primary release date on or after `{year}-01-01`,
        and primary release date on or before six months before today.

        Results are sorted by `primary_release_date.desc`. Pages are fetched
        sequentially up to TMDB's reported `total_pages`, capped by the server
        (`tmdbMaxDiscoverPages`; TMDB itself also caps Discover at 500 pages).

        Each page is upserted on `tmdb_id`. The `saved` count includes both
        inserts and updates. This call can take a long time when many pages
        are available.
      parameters:
        - name: year
          in: query
          required: true
          description: |
            Inclusive lower bound for primary release year. Must be an integer
            from 1870 through the current calendar year plus one.
          schema:
            type: integer
            minimum: 1870
            example: 2020
      responses:
        "200":
          description: Ingest finished.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PopulateMoviesSinceResponse"
        "400":
          description: Missing or invalid `year` query parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "502":
          description: TMDB request failed while fetching a page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PopulateMoviesSinceError"
        "500":
          description: TMDB payload could not be parsed or movies could not be saved.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/Error"
                  - $ref: "#/components/schemas/PopulateMoviesSinceError"
        default:
          description: TMDB returned a non-200 status while fetching a page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PopulateMoviesSinceError"

components:
  schemas:
    Movie:
      type: object
      additionalProperties: false
      required:
        - id
        - tmdb_id
        - title
        - original_title
        - overview
        - poster_path
        - backdrop_path
        - original_language
        - vote_average
        - vote_count
        - popularity
        - adult
        - genre_ids
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: Local primary key.
          example: 405
        tmdb_id:
          type: integer
          description: TMDB movie id. Unique in this catalog.
          example: 1204680
        title:
          type: string
          example: Coyote vs. Acme
        original_title:
          type: string
          example: Coyote vs. Acme
        overview:
          type: string
        release_date:
          type: string
          format: date-time
          nullable: true
          description: Primary release date, RFC 3339. Null when TMDB had no parseable date.
        poster_path:
          type: string
          description: TMDB poster path (not a full URL).
          example: /vhv7lBWYM0DUuNU2a0V7Rhq21dD.jpg
        backdrop_path:
          type: string
          description: TMDB backdrop path (not a full URL).
        original_language:
          type: string
          example: en
        vote_average:
          type: number
          format: float
        vote_count:
          type: integer
        popularity:
          type: number
          format: float
        adult:
          type: boolean
        genre_ids:
          type: array
          items:
            type: integer
          description: TMDB genre ids.
          example: [35, 12, 10751]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    Genre:
      type: object
      additionalProperties: false
      required:
        - id
        - media_type
        - name
      properties:
        id:
          type: integer
          description: TMDB genre id.
          example: 35
        media_type:
          type: string
          enum: [movie, tv]
          description: Which TMDB genre list this row came from. Part of the composite primary key alongside id.
          example: movie
        name:
          type: string
          example: Comedy

    PopulateGenresResponse:
      type: object
      additionalProperties: false
      required: [saved, genres]
      properties:
        saved:
          type: integer
          description: Number of genres upserted from this request.
        genres:
          type: array
          items:
            $ref: "#/components/schemas/Genre"

    ListGenresResponse:
      type: object
      additionalProperties: false
      required: [genres]
      properties:
        genres:
          type: array
          items:
            $ref: "#/components/schemas/Genre"

    ListMoviesResponse:
      type: object
      additionalProperties: false
      required: [page, page_size, total_pages, total_results, movies]
      properties:
        page:
          type: integer
          description: 1-based page that was returned.
          example: 1
        page_size:
          type: integer
          description: Number of movies requested per page.
          example: 20
        total_pages:
          type: integer
          description: Total pages at this page_size. 0 when the catalog is empty.
        total_results:
          type: integer
          description: Total non-deleted movies in the catalog.
        movies:
          type: array
          items:
            $ref: "#/components/schemas/Movie"

    PopulateMoviesResponse:
      type: object
      additionalProperties: false
      required: [saved, movies]
      properties:
        saved:
          type: integer
          description: Number of movies upserted from this request.
        movies:
          type: array
          items:
            $ref: "#/components/schemas/Movie"

    PopulateMoviesSinceResponse:
      type: object
      additionalProperties: false
      required: [year, pages_fetched, total_pages, saved]
      properties:
        year:
          type: integer
          description: The `year` query value that was ingested.
        pages_fetched:
          type: integer
          description: How many TMDB pages were successfully fetched.
        total_pages:
          type: integer
          description: Page count used for the loop (TMDB total, after server cap).
        saved:
          type: integer
          description: Movies upserted across all fetched pages (inserts plus updates).

    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
          example: failed to load movies

    PopulateMoviesSinceError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
        year:
          type: integer
        pages_fetched:
          type: integer
        saved:
          type: integer